implement start_game -> score_game pipeline for Mistral-7B-Instruct-v0.1
This commit adds a new pipeline. It copies a lot of stuff from the
chatgpt_pipeline. The only difference is the implementation of `start_game`
which loads Mistral-7B-Instruct-v0.1 into memory and calls it with the prompts.

Note that mistral does not support a "system" prompt the way GPT-3.5 does.
This pipeline approximates the same behavior by prepending the system prompt
to the user prompt. However, this might be resulting in artificially low
scores for mistral, as the system prompt is not included when we send the
response to GPT-4 for scoring.