implement pipeline with strategy pattern
This commit refactors the `pipeline` file so that it offers an `evaluate`
method which accepts a `start_game` implementation. This method does
everything you need to invoke a particular model / prompt and then apply
judging to it. You can add new implementations which simply implement the
model invocation and easily judge them by calling `evaluate`.