initial D&D 5e benchmark (#1)
This benchmark set is intended to cover fine-tuning for
an agent who can make decisions about how D&D 5th edition
rules as written might apply to specific scenarios.

This implements a basic pipeline which, given a question,
a) generates keywords
b) queries ElasticSearch using a combination of the keywords and KNN similarity search on the question itself (ingestion into Elastic happens outside this repo)
c) passes the resulting documents as context to a new prompt
d) records the results

All results are written out to a CSV file for now. This is the basic implementation of a benchmark we will use for different models, parameterizations, search strategies, prompts, etc.